pyproject.toml
[tool.poetry]
name = "google-trend-bot"
version = "0.1.0"
description = ""
authors = ["humaningansalam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.10.*"
pytz = "^2024.2"
requests = "^2.32.3"
schedule = "^1.2.2"
selenium = "^4.25.0"
fluent-logger = "^0.11.1"
prometheus-client = "^0.21.0"
feedparser = "^6.0.11"
flask = "^3.0.3"
psutil = "^6.1.0"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
